refactor(models): extract domain models to app/models/
Move SagData, SurfaceData, Wavelength, OpticalSystemData, TaskSource,
ConnectionState, MaxPVResult from dde/types.h to app/models/.

dde/types.h becomes a thin re-export layer using declarations for
backward compatibility. initial_data_load_service.h now includes types.h
instead of forward-declaring OpticalSystemData.
